+++ src/parallel_cv.hpp
@@ -5,7 +5,7 @@
 
 namespace parallel_cv {
   void run(cv::String video_path,
-           int num_worker_threads,
+           size_t num_worker_threads,
            cv::Mat (*const work_function)(cv::Mat&, cv::Mat&));
 }
